+++ NRedisTimeSeries/TimeSeriesClientAsync.cs
@@ -128,10 +128,10 @@ namespace NRedisTimeSeries
         /// <param name="fromTimeStamp">Start timestamp for the range deletion.</param>
         /// <param name="toTimeStamp">End timestamp for the range deletion.</param>
         /// <returns>The count of deleted items</returns>
-        public static async Task<int> TimeSeriesDelAsync(this IDatabase db, string key, TimeStamp fromTimeStamp, TimeStamp toTimeStamp) 
+        public static async Task<long> TimeSeriesDelAsync(this IDatabase db, string key, TimeStamp fromTimeStamp, TimeStamp toTimeStamp) 
         {
             var args = BuildTsDelArgs(key, fromTimeStamp, toTimeStamp);
-            return ParseInteger(await db.ExecuteAsync(TS.DEL, args));
+            return ParseLong(await db.ExecuteAsync(TS.DEL, args));
         }
 
         #endregion
